+++ src/app/app.component.ts
@@ -189,7 +189,7 @@ export class AppComponent {
           || this.userService.IsLocalManager()
         }, {
           name: 'Report an Issue',
-          href: `mailto:reportingportal.help@jira.itransition.com?body=${this.issueEmailBody}`,
+          href: `https://github.com/aquality-automation/aquality-tracking/issues`,
           show: true
         }]
       }];