+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 0.3.2 (unreleased)
 
-Features
+Features:
 
+  - Report an issue now redirecting to Github -> [View Issue](https://github.com/aquality-automation/aquality-tracking-ui/issues/28)
   - Import: NUnit V3 add possibility to import with fullname test name  -> [View Issue](https://github.com/aquality-automation/aquality-tracking-ui/issues/27)
   
 Bugfixes: